refactor: layout_parser.rs を責務別の4サブモジュールに分割 (#17)
- field_extractor.rs: フィールド参照・Bounds・ラベル抽出
- script_and_format.rs: スクリプトトリガー・条件付き書式
- object_scanner.rs: ObjectList/Object 走査（中核）
- layout_catalog.rs: parse_layouts エントリーポイント + 全テスト(35件)
