+++ src-tauri/src/parser/layout_parser/layout_catalog.rs
@@ -3,20 +3,27 @@ use std::io::BufRead;
 use quick_xml::{events::Event, Reader};
 
 use crate::parser::{
-    helpers::{get_attr, read_text_content, skip_element},
-    models::{
-        Bounds, ConditionalFormat, Layout, LayoutFieldRef, LayoutId, LayoutObject, ScriptTrigger,
-    },
+    helpers::{get_attr, skip_element},
+    models::{Layout, LayoutFieldRef, LayoutId, LayoutObject, ScriptTrigger},
     ParseError,
 };
 
-/// `scan_object` / `scan_object_list` の戻り値型エイリアス。
-type ObjectScanResult = (
-    Vec<ScriptTrigger>,
-    Vec<String>,
-    Vec<LayoutFieldRef>,
-    Vec<LayoutObject>,
-);
+use super::{
+    object_scanner::{scan_object, scan_object_list},
+    script_and_format::parse_script_trigger_elements,
+};
+
+/// `parse_layout_children` の戻り値型エイリアス。
+type LayoutChildrenResult = Result<
+    (
+        Option<String>,
+        Vec<ScriptTrigger>,
+        Vec<String>,
+        Vec<LayoutFieldRef>,
+        Vec<LayoutObject>,
+    ),
+    ParseError,
+>;
 
 /// Parse `<LayoutCatalog>` content.
 ///
@@ -98,18 +105,6 @@ pub fn parse_layouts<R: BufRead>(
 /// Parse children of a `<Layout>` element (opening tag already consumed).
 ///
 /// Returns `(table_occurrence_name, triggers, button_script_refs, field_refs, layout_objects)`.
-/// `parse_layout_children` の戻り値型エイリアス。
-type LayoutChildrenResult = Result<
-    (
-        Option<String>,
-        Vec<ScriptTrigger>,
-        Vec<String>,
-        Vec<LayoutFieldRef>,
-        Vec<LayoutObject>,
-    ),
-    ParseError,
->;
-
 fn parse_layout_children<R: BufRead>(
     reader: &mut Reader<R>,
     buf: &mut Vec<u8>,
@@ -192,670 +187,6 @@ fn parse_layout_children<R: BufRead>(
         field_refs,
         layout_objects,
     ))
-}
-
-/// `<ObjectList>` の内容を走査してトリガー・ボタンスクリプト参照・フィールド参照・オブジェクト一覧を収集する。
-fn scan_object_list<R: BufRead>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-) -> Result<ObjectScanResult, ParseError> {
-    let mut triggers = Vec::new();
-    let mut button_scripts = Vec::new();
-    let mut field_refs = Vec::new();
-    let mut layout_objects = Vec::new();
-
-    loop {
-        buf.clear();
-        match reader.read_event_into(buf)? {
-            Event::Start(ref e) if e.name().as_ref() == b"Object" => {
-                let obj_type = get_attr(e, b"type").unwrap_or_default();
-                let obj_key = get_attr(e, b"key")
-                    .and_then(|v| {
-                        v.parse::<u64>()
-                            .map_err(|_| ParseError::InvalidValue(format!("Object key: {v}")))
-                    })
-                    .unwrap_or(0);
-                let obj_name = get_attr(e, b"name").ok().filter(|s| !s.is_empty());
-                let (ot, os, ofr, olo) = scan_object(reader, buf, obj_type, obj_key, obj_name)?;
-                triggers.extend(ot);
-                button_scripts.extend(os);
-                field_refs.extend(ofr);
-                layout_objects.extend(olo);
-            }
-            Event::Empty(ref e) if e.name().as_ref() == b"Object" => {
-                let obj_type = get_attr(e, b"type").unwrap_or_default();
-                let obj_key = get_attr(e, b"key")
-                    .and_then(|v| {
-                        v.parse::<u64>()
-                            .map_err(|_| ParseError::InvalidValue(format!("Object key: {v}")))
-                    })
-                    .unwrap_or(0);
-                let obj_name = get_attr(e, b"name").ok().filter(|s| !s.is_empty());
-                layout_objects.push(LayoutObject {
-                    object_type: obj_type,
-                    object_key: obj_key,
-                    object_name: obj_name,
-                    button_label: None,
-                    field_table_occurrence: None,
-                    field_name: None,
-                    tooltip: None,
-                    hide_condition: None,
-                    bounds: None,
-                    conditional_formats: Vec::new(),
-                });
-            }
-            Event::Start(_) => {
-                skip_element(reader, buf)?;
-            }
-            Event::Empty(_) => {}
-            Event::End(_) => break, // </ObjectList>
-            Event::Eof => return Err(ParseError::UnexpectedEof),
-            _ => {}
-        }
-    }
-
-    Ok((triggers, button_scripts, field_refs, layout_objects))
-}
-
-/// `<FieldReference>` 要素から `(tableOccurrence, field)` を抽出する。
-fn parse_field_reference_attrs(e: &quick_xml::events::BytesStart<'_>) -> Option<LayoutFieldRef> {
-    let table_occurrence = get_attr(e, b"tableOccurrence")
-        .or_else(|_| get_attr(e, b"table"))
-        .unwrap_or_default();
-    let field_name = get_attr(e, b"field").unwrap_or_default();
-    if !table_occurrence.is_empty() && !field_name.is_empty() {
-        Some(LayoutFieldRef {
-            table_occurrence,
-            field_name,
-        })
-    } else {
-        None
-    }
-}
-
-/// `<Field name="..." table="..."/>` 要素からフィールド参照を抽出する。
-fn parse_field_elem_attrs(e: &quick_xml::events::BytesStart<'_>) -> Option<LayoutFieldRef> {
-    let table_occurrence = get_attr(e, b"table").unwrap_or_default();
-    let field_name = get_attr(e, b"name").unwrap_or_default();
-    if !table_occurrence.is_empty() && !field_name.is_empty() {
-        Some(LayoutFieldRef {
-            table_occurrence,
-            field_name,
-        })
-    } else {
-        None
-    }
-}
-
-/// `<ToolTip>` や `<HideCondition>` の子要素 `<Calculation>` から計算式文字列を取得する。
-///
-/// 呼び出し時点: 親の開始タグ直後。親の閉じタグまで消費して返す。
-fn read_calculation_from_container<R: BufRead>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-) -> Result<Option<String>, ParseError> {
-    let mut result: Option<String> = None;
-
-    loop {
-        buf.clear();
-        match reader.read_event_into(buf)? {
-            Event::Start(ref e) if e.name().as_ref() == b"Calculation" => {
-                let text = read_text_content(reader, buf)?;
-                if !text.is_empty() {
-                    result = Some(text);
-                }
-            }
-            Event::Empty(ref e) if e.name().as_ref() == b"Calculation" => {}
-            Event::Start(_) => {
-                skip_element(reader, buf)?;
-            }
-            Event::Empty(_) => {}
-            Event::End(_) => break,
-            Event::Eof => return Err(ParseError::UnexpectedEof),
-            _ => {}
-        }
-    }
-
-    Ok(result)
-}
-
-/// `<FieldObj>` 内を走査して `<DDRInfo><Field name="..." table="..."/>` を返す。
-fn scan_field_obj<R: BufRead>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-) -> Result<(Option<String>, Option<String>), ParseError> {
-    let mut table_occ: Option<String> = None;
-    let mut field_name: Option<String> = None;
-
-    loop {
-        buf.clear();
-        match reader.read_event_into(buf)? {
-            Event::Start(ref e) if e.name().as_ref() == b"DDRInfo" => {
-                let (t, f) = scan_ddr_info(reader, buf)?;
-                table_occ = t;
-                field_name = f;
-            }
-            Event::Empty(ref e) if e.name().as_ref() == b"DDRInfo" => {}
-            Event::Start(_) => {
-                skip_element(reader, buf)?;
-            }
-            Event::Empty(_) => {}
-            Event::End(_) => break, // </FieldObj>
-            Event::Eof => return Err(ParseError::UnexpectedEof),
-            _ => {}
-        }
-    }
-    Ok((table_occ, field_name))
-}
-
-/// `<DDRInfo>` 内を走査して `<Field name="..." table="..."/>` を返す。
-fn scan_ddr_info<R: BufRead>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-) -> Result<(Option<String>, Option<String>), ParseError> {
-    let mut table_occ: Option<String> = None;
-    let mut field_name: Option<String> = None;
-
-    loop {
-        buf.clear();
-        match reader.read_event_into(buf)? {
-            Event::Empty(ref e) if e.name().as_ref() == b"Field" => {
-                table_occ = get_attr(e, b"table").ok();
-                field_name = get_attr(e, b"name").ok();
-            }
-            Event::Start(ref e) if e.name().as_ref() == b"Field" => {
-                table_occ = get_attr(e, b"table").ok();
-                field_name = get_attr(e, b"name").ok();
-                skip_element(reader, buf)?;
-            }
-            Event::Start(_) => {
-                skip_element(reader, buf)?;
-            }
-            Event::Empty(_) => {}
-            Event::End(_) => break, // </DDRInfo>
-            Event::Eof => return Err(ParseError::UnexpectedEof),
-            _ => {}
-        }
-    }
-    Ok((table_occ, field_name))
-}
-
-/// `<Object>` 要素内を走査して ScriptTrigger・ボタンスクリプト参照・フィールド参照・LayoutObject を収集する。
-///
-/// `object_type` と `object_key` は呼び出し元が `<Object>` 開始タグから読み取って渡す。
-fn scan_object<R: BufRead>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    object_type: String,
-    object_key: u64,
-    object_name: Option<String>,
-) -> Result<ObjectScanResult, ParseError> {
-    let mut triggers = Vec::new();
-    let mut button_scripts = Vec::new();
-    let mut field_refs: Vec<LayoutFieldRef> = Vec::new();
-    let mut layout_objects: Vec<LayoutObject> = Vec::new();
-
-    // このオブジェクト自身の属性
-    let mut field_table_occurrence: Option<String> = None;
-    let mut field_name_val: Option<String> = None;
-    let mut tooltip: Option<String> = None;
-    let mut hide_condition: Option<String> = None;
-    let mut bounds: Option<Bounds> = None;
-    let mut conditional_formats: Vec<ConditionalFormat> = Vec::new();
-
-    let is_field_obj = object_type == "Field";
-    let mut button_label: Option<String> = None;
-
-    loop {
-        buf.clear();
-        match reader.read_event_into(buf)? {
-            // ---- Bounds（位置・サイズ情報）----
-            Event::Empty(ref e) if e.name().as_ref() == b"Bounds" => {
-                bounds = Some(parse_bounds(e));
-            }
-
-            // ---- ToolTip ----
-            Event::Start(ref e) if e.name().as_ref() == b"ToolTip" => {
-                tooltip = read_calculation_from_container(reader, buf)?;
-            }
-            Event::Empty(ref e) if e.name().as_ref() == b"ToolTip" => {}
-
-            // ---- HideCondition ----
-            Event::Start(ref e) if e.name().as_ref() == b"HideCondition" => {
-                hide_condition = read_calculation_from_container(reader, buf)?;
-            }
-            Event::Empty(ref e) if e.name().as_ref() == b"HideCondition" => {}
-
-            // ---- FieldObj（type="Field" のみ解析）----
-            Event::Start(ref e) if e.name().as_ref() == b"FieldObj" => {
-                if is_field_obj {
-                    let (ft, fn_) = scan_field_obj(reader, buf)?;
-                    field_table_occurrence = ft.clone();
-                    field_name_val = fn_.clone();
-                    // layout_field_refs 用にも追加
-                    if let (Some(t), Some(f)) = (ft, fn_) {
-                        field_refs.push(LayoutFieldRef {
-                            table_occurrence: t,
-                            field_name: f,
-                        });
-                    }
-                } else {
-                    let (nested_scripts, nested_frs) =
-                        deep_scan_for_scripts_and_fields(reader, buf)?;
-                    button_scripts.extend(nested_scripts);
-                    field_refs.extend(nested_frs);
-                }
-            }
-            Event::Empty(ref e) if e.name().as_ref() == b"FieldObj" => {}
-
-            // ---- FieldReference (self-closing) ----
-            Event::Empty(ref e) if e.name().as_ref() == b"FieldReference" => {
-                if let Some(fr) = parse_field_reference_attrs(e) {
-                    field_refs.push(fr);
-                }
-            }
-            // ---- FieldReference (with children) ----
-            Event::Start(ref e) if e.name().as_ref() == b"FieldReference" => {
-                if let Some(fr) = parse_field_reference_attrs(e) {
-                    field_refs.push(fr);
-                }
-                skip_element(reader, buf)?;
-            }
-            // ---- 実DDR形式: <Field name="..." table="..."/> (self-closing) ----
-            Event::Empty(ref e) if e.name().as_ref() == b"Field" => {
-                if let Some(fr) = parse_field_elem_attrs(e) {
-                    field_refs.push(fr);
-                }
-            }
-            // ---- 実DDR形式: <Field name="..." table="..."> (with children) ----
-            Event::Start(ref e) if e.name().as_ref() == b"Field" => {
-                if let Some(fr) = parse_field_elem_attrs(e) {
-                    field_refs.push(fr);
-                }
-                let (_, nested_frs) = deep_scan_for_scripts_and_fields(reader, buf)?;
-                field_refs.extend(nested_frs);
-            }
-
-            // ---- ScriptTriggerList ----
-            Event::Start(ref e) if e.name().as_ref() == b"ScriptTriggerList" => {
-                triggers.extend(parse_script_trigger_elements(
-                    reader,
-                    buf,
-                    b"ScriptTrigger",
-                )?);
-            }
-            Event::Empty(ref e) if e.name().as_ref() == b"ScriptTriggerList" => {}
-            // ---- 旧形式: ScriptTriggers ----
-            Event::Start(ref e) if e.name().as_ref() == b"ScriptTriggers" => {
-                triggers.extend(parse_script_trigger_elements(reader, buf, b"Trigger")?);
-            }
-            Event::Empty(ref e) if e.name().as_ref() == b"ScriptTriggers" => {}
-
-            // ---- ネストされた Object ----
-            Event::Start(ref e) if e.name().as_ref() == b"Object" => {
-                let nested_type = get_attr(e, b"type").unwrap_or_default();
-                let nested_key = get_attr(e, b"key")
-                    .and_then(|v| {
-                        v.parse::<u64>()
-                            .map_err(|_| ParseError::InvalidValue(format!("Object key: {v}")))
-                    })
-                    .unwrap_or(0);
-                let nested_name = get_attr(e, b"name").ok().filter(|s| !s.is_empty());
-                let (nt, ns, nfr, nlo) =
-                    scan_object(reader, buf, nested_type, nested_key, nested_name)?;
-                triggers.extend(nt);
-                button_scripts.extend(ns);
-                field_refs.extend(nfr);
-                layout_objects.extend(nlo);
-            }
-            // ---- ネストされた ObjectList ----
-            Event::Start(ref e) if e.name().as_ref() == b"ObjectList" => {
-                let (nt, ns, nfr, nlo) = scan_object_list(reader, buf)?;
-                triggers.extend(nt);
-                button_scripts.extend(ns);
-                field_refs.extend(nfr);
-                layout_objects.extend(nlo);
-            }
-            Event::Empty(ref e) if e.name().as_ref() == b"ObjectList" => {}
-
-            // ---- Script (self-closing) ----
-            Event::Empty(ref e) if e.name().as_ref() == b"Script" => {
-                if let Ok(name) = get_attr(e, b"name") {
-                    if !name.is_empty() {
-                        button_scripts.push(name);
-                    }
-                }
-            }
-            // ---- Script (with children) ----
-            Event::Start(ref e) if e.name().as_ref() == b"Script" => {
-                if let Ok(name) = get_attr(e, b"name") {
-                    if !name.is_empty() {
-                        button_scripts.push(name);
-                    }
-                }
-                skip_element(reader, buf)?;
-            }
-            // ---- ScriptReference (self-closing) ----
-            Event::Empty(ref e) if e.name().as_ref() == b"ScriptReference" => {
-                if let Ok(name) = get_attr(e, b"name") {
-                    if !name.is_empty() {
-                        button_scripts.push(name);
-                    }
-                }
-            }
-            // ---- TextObj（ボタンラベル取得）----
-            Event::Start(ref e) if e.name().as_ref() == b"TextObj" => {
-                button_label = read_text_obj_label(reader, buf)?;
-            }
-            Event::Empty(ref e) if e.name().as_ref() == b"TextObj" => {}
-
-            // ---- ConditionalFormatting ----
-            Event::Start(ref e) if e.name().as_ref() == b"ConditionalFormatting" => {
-                conditional_formats = parse_conditional_formatting(reader, buf)?;
-            }
-            Event::Empty(ref e) if e.name().as_ref() == b"ConditionalFormatting" => {}
-
-            // その他の子要素は深掘りスキャン
-            Event::Start(_) => {
-                let (nested_scripts, nested_frs) = deep_scan_for_scripts_and_fields(reader, buf)?;
-                button_scripts.extend(nested_scripts);
-                field_refs.extend(nested_frs);
-            }
-            Event::Empty(_) => {}
-            Event::End(_) => {
-                // このオブジェクト自身の LayoutObject を追加
-                layout_objects.push(LayoutObject {
-                    object_type,
-                    object_key,
-                    object_name,
-                    button_label,
-                    field_table_occurrence,
-                    field_name: field_name_val,
-                    tooltip,
-                    hide_condition,
-                    bounds,
-                    conditional_formats,
-                });
-                break;
-            }
-            Event::Eof => return Err(ParseError::UnexpectedEof),
-            _ => {}
-        }
-    }
-
-    Ok((triggers, button_scripts, field_refs, layout_objects))
-}
-
-/// `<TextObj>` 内の最初の `<Data>` テキストを取り出す（ボタンラベル取得用）。
-/// 呼び出し時点で `<TextObj>` の開始タグは消費済み。
-fn read_text_obj_label<R: BufRead>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-) -> Result<Option<String>, ParseError> {
-    let mut depth: u32 = 1;
-    let mut in_data = false;
-    let mut label: Option<String> = None;
-
-    loop {
-        buf.clear();
-        match reader.read_event_into(buf)? {
-            Event::Start(ref e) => {
-                if e.name().as_ref() == b"Data" && label.is_none() {
-                    in_data = true;
-                }
-                depth += 1;
-            }
-            Event::Text(ref e) if in_data => {
-                if let Ok(decoded) = e.decode() {
-                    let text = quick_xml::escape::unescape(&decoded)
-                        .unwrap_or_default()
-                        .trim()
-                        .to_string();
-                    if !text.is_empty() {
-                        label = Some(text);
-                    }
-                }
-                in_data = false;
-            }
-            Event::End(_) => {
-                depth -= 1;
-                if depth == 0 {
-                    break;
-                }
-                in_data = false;
-            }
-            Event::Eof => return Err(ParseError::UnexpectedEof),
-            _ => {
-                in_data = false;
-            }
-        }
-    }
-
-    Ok(label)
-}
-
-/// 任意の要素内を深掘りして、スクリプト参照とフィールド参照を両方収集する。
-fn deep_scan_for_scripts_and_fields<R: BufRead>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-) -> Result<(Vec<String>, Vec<LayoutFieldRef>), ParseError> {
-    let mut scripts = Vec::new();
-    let mut field_refs = Vec::new();
-    let mut depth: u32 = 1;
-
-    loop {
-        buf.clear();
-        match reader.read_event_into(buf)? {
-            Event::Start(ref e) => {
-                let name_bytes = e.name().as_ref().to_vec();
-                if name_bytes == b"Script" || name_bytes == b"ScriptReference" {
-                    let file = get_attr(e, b"file").unwrap_or_default();
-                    if file.is_empty() {
-                        if let Ok(name) = get_attr(e, b"name") {
-                            if !name.is_empty() {
-                                scripts.push(name);
-                            }
-                        }
-                    }
-                } else if name_bytes == b"FieldReference" {
-                    if let Some(fr) = parse_field_reference_attrs(e) {
-                        field_refs.push(fr);
-                    }
-                } else if name_bytes == b"Field" {
-                    if let Some(fr) = parse_field_elem_attrs(e) {
-                        field_refs.push(fr);
-                    }
-                }
-                depth += 1;
-            }
-            Event::Empty(ref e) => {
-                let name_bytes = e.name().as_ref().to_vec();
-                if name_bytes == b"Script" || name_bytes == b"ScriptReference" {
-                    let file = get_attr(e, b"file").unwrap_or_default();
-                    if file.is_empty() {
-                        if let Ok(name) = get_attr(e, b"name") {
-                            if !name.is_empty() {
-                                scripts.push(name);
-                            }
-                        }
-                    }
-                } else if name_bytes == b"FieldReference" {
-                    if let Some(fr) = parse_field_reference_attrs(e) {
-                        field_refs.push(fr);
-                    }
-                } else if name_bytes == b"Field" {
-                    if let Some(fr) = parse_field_elem_attrs(e) {
-                        field_refs.push(fr);
-                    }
-                }
-            }
-            Event::End(_) => {
-                depth -= 1;
-                if depth == 0 {
-                    break;
-                }
-            }
-            Event::Eof => return Err(ParseError::UnexpectedEof),
-            _ => {}
-        }
-    }
-
-    Ok((scripts, field_refs))
-}
-
-/// トリガーコンテナ（`<ScriptTriggerList>` または `<ScriptTriggers>`）の内容をパース。
-fn parse_script_trigger_elements<R: BufRead>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-    trigger_elem: &[u8],
-) -> Result<Vec<ScriptTrigger>, ParseError> {
-    let mut triggers = Vec::new();
-
-    loop {
-        buf.clear();
-        match reader.read_event_into(buf)? {
-            Event::Start(ref e) if e.name().as_ref() == trigger_elem => {
-                let event = get_attr(e, b"event").unwrap_or_default();
-                let (script_name, file_name) = parse_trigger_script_ref(reader, buf)?;
-                if !script_name.is_empty() {
-                    triggers.push(ScriptTrigger {
-                        event,
-                        script_name,
-                        file_name,
-                    });
-                }
-            }
-            Event::Empty(ref e) if e.name().as_ref() == trigger_elem => {}
-            Event::Start(_) => {
-                skip_element(reader, buf)?;
-            }
-            Event::Empty(_) => {}
-            Event::End(_) => break,
-            Event::Eof => return Err(ParseError::UnexpectedEof),
-            _ => {}
-        }
-    }
-
-    Ok(triggers)
-}
-
-/// `<ScriptTrigger>` または `<Trigger>` の子要素から `<Script name="..." file="..."/>` を取得する。
-fn parse_trigger_script_ref<R: BufRead>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-) -> Result<(String, String), ParseError> {
-    let mut script_name = String::new();
-    let mut file_name = String::new();
-
-    loop {
-        buf.clear();
-        match reader.read_event_into(buf)? {
-            Event::Empty(ref e) if e.name().as_ref() == b"Script" => {
-                script_name = get_attr(e, b"name").unwrap_or_default();
-                file_name = get_attr(e, b"file").unwrap_or_default();
-            }
-            Event::Start(ref e) if e.name().as_ref() == b"Script" => {
-                script_name = get_attr(e, b"name").unwrap_or_default();
-                file_name = get_attr(e, b"file").unwrap_or_default();
-                skip_element(reader, buf)?;
-            }
-            Event::Start(_) => {
-                skip_element(reader, buf)?;
-            }
-            Event::Empty(_) => {}
-            Event::End(_) => break,
-            Event::Eof => return Err(ParseError::UnexpectedEof),
-            _ => {}
-        }
-    }
-
-    Ok((script_name, file_name))
-}
-
-/// `<Bounds top="..." left="..." bottom="..." right="..."/>` 属性から `Bounds` を生成する。
-fn parse_bounds(e: &quick_xml::events::BytesStart<'_>) -> Bounds {
-    let mut top = 0.0f64;
-    let mut left = 0.0f64;
-    let mut bottom = 0.0f64;
-    let mut right = 0.0f64;
-    for attr in e.attributes().flatten() {
-        let val: f64 = std::str::from_utf8(attr.value.as_ref())
-            .ok()
-            .and_then(|s| s.parse().ok())
-            .unwrap_or(0.0);
-        match attr.key.as_ref() {
-            b"top" => top = val,
-            b"left" => left = val,
-            b"bottom" => bottom = val,
-            b"right" => right = val,
-            _ => {}
-        }
-    }
-    Bounds {
-        top,
-        left,
-        bottom,
-        right,
-    }
-}
-
-/// `<ConditionalFormatting>` ブロックをパースして `ConditionalFormat` のベクタを返す。
-/// 呼び出し時点で `<ConditionalFormatting>` 開始タグは消費済み。
-fn parse_conditional_formatting<R: BufRead>(
-    reader: &mut Reader<R>,
-    buf: &mut Vec<u8>,
-) -> Result<Vec<ConditionalFormat>, ParseError> {
-    let mut formats: Vec<ConditionalFormat> = Vec::new();
-    let mut current_rule_order: u32 = 0;
-    let mut current_calculation: Option<String> = None;
-    let mut current_format_css: Option<String> = None;
-
-    loop {
-        buf.clear();
-        match reader.read_event_into(buf)? {
-            Event::Start(ref e) => match e.name().as_ref() {
-                b"Item" => {
-                    current_rule_order = get_attr(e, b"id")
-                        .ok()
-                        .and_then(|v| v.parse::<u32>().ok())
-                        .unwrap_or(0);
-                    current_calculation = None;
-                    current_format_css = None;
-                }
-                b"Calculation" => {
-                    let text = read_text_content(reader, buf)?;
-                    if !text.is_empty() {
-                        current_calculation = Some(text);
-                    }
-                }
-                b"LocalCSS" => {
-                    let text = read_text_content(reader, buf)?;
-                    if !text.is_empty() {
-                        current_format_css = Some(text);
-                    }
-                }
-                _ => {}
-            },
-            Event::End(ref e) => match e.name().as_ref() {
-                b"ConditionalFormatting" => break,
-                b"Item" => {
-                    if let Some(calc) = current_calculation.take() {
-                        formats.push(ConditionalFormat {
-                            rule_order: current_rule_order,
-                            calculation: calc,
-                            format_css: current_format_css.take().unwrap_or_default(),
-                        });
-                    }
-                }
-                _ => {}
-            },
-            Event::Empty(_) => {}
-            Event::Eof => return Err(ParseError::UnexpectedEof),
-            _ => {}
-        }
-    }
-
-    Ok(formats)
 }
 
 // ---------------------------------------------------------------------------
